pyproject.toml
[project]
name = "antdx-demo-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
